fixing bugs with package module
